feat(gui-v2): use separate api instances with `useApi`
